+++ src/AlpacaDevice.h
@@ -63,7 +63,7 @@ protected:
     void AlpacaGetSupportedActions(AsyncWebServerRequest *request);
 
     // helpers
-    int32_t checkClientDataAndConnection(AsyncWebServerRequest *request, uint32_t &clientIdx, Spelling_t spelling);
+    int32_t checkClientDataAndConnection(AsyncWebServerRequest *request, uint32_t &clientIdx, Spelling_t spelling, bool check_connection = true );
     uint32_t getClientIdxByClientID(uint32_t clientID);
 
 public: